pyproject.toml
[project]
name = "binance-grid-bot"
version = "1.0.0"
description = "AI-Enhanced Cryptocurrency Grid Trading Bot"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "murriiii", email = "murriiii@users.noreply.github.com" }
]
keywords = ["trading", "cryptocurrency", "binance", "grid-trading", "bot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "python-binance>=1.0.17",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "schedule>=1.2.0",
    "python-telegram-bot>=20.0",
    "matplotlib>=3.7.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[project.urls]
Homepage = "https://github.com/murriiii/binance-grid-bot"
Repository = "https://github.com/murriiii/binance-grid-bot.git"
Issues = "https://github.com/murriiii/binance-grid-bot/issues"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

# ═══════════════════════════════════════════════════════════════
# RUFF - Fast Python Linter & Formatter
# ═══════════════════════════════════════════════════════════════
[tool.ruff]
target-version = "py310"
line-length = 100
indent-width = 4

# Include/Exclude
include = ["*.py", "*.pyi"]
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".mypy_cache",
    ".pytest_cache",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable rules
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]

# Ignore specific rules
ignore = [
    "E501",    # Line too long (handled by formatter)
    "B008",    # Do not perform function calls in argument defaults
    "PLR0913", # Too many arguments
    "PLR2004", # Magic value used in comparison
    "PLR0912", # Too many branches (complex functions are acceptable)
    "PLR0915", # Too many statements (complex functions are acceptable)
    "PLR0911", # Too many return statements
    "PLC0415", # Import at top-level (optional imports in try blocks)
    "PLW0603", # Global statement (singleton pattern)
    "PLW0108", # Lambda may be unnecessary
    "ERA001",  # Commented-out code (documentation comments)
    "RUF012",  # Mutable class attributes (ClassVar)
    "RUF013",  # Implicit Optional
    "RUF059",  # Unpacked variable not used
    "PTH123",  # open() vs Path.open()
    "B007",    # Unused loop control variable
    "B905",    # zip() without strict parameter
    "C408",    # Unnecessary dict() call
    "PLW1508", # Invalid type for env var default (int/float)
    "F841",    # Unused local variable (intentional sometimes)
    "E402",    # Module level import not at top (load_dotenv pattern)
    "RUF001",  # Ambiguous unicode characters (intentional emojis)
    "RUF010",  # Use explicit conversion flag
    "ARG001",  # Unused function argument (often needed for callbacks)
    "ARG002",  # Unused method argument
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["src"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004"]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = 80

# ═══════════════════════════════════════════════════════════════
# MYPY - Type Checking
# ═══════════════════════════════════════════════════════════════
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "binance.*",
    "psycopg2.*",
    "telegram.*",
    "schedule.*",
]
ignore_missing_imports = true

# ═══════════════════════════════════════════════════════════════
# PYTEST
# ═══════════════════════════════════════════════════════════════
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--tb=short",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

# ═══════════════════════════════════════════════════════════════
# COVERAGE
# ═══════════════════════════════════════════════════════════════
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
fail_under = 60

# ═══════════════════════════════════════════════════════════════
# COMMITIZEN - Conventional Commits
# ═══════════════════════════════════════════════════════════════
[tool.commitizen]
name = "cz_conventional_commits"
version = "1.0.0"
version_files = [
    "pyproject.toml:version",
    "src/__init__.py:__version__",
]
tag_format = "v$version"
update_changelog_on_bump = true
changelog_file = "CHANGELOG.md"
